% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{autolayer.Assay}
\alias{autolayer.Assay}
\title{Create a Layer to Add to \code{ggplot} Objects}
\usage{
\method{autolayer}{Assay}(
  object,
  data = missing_arg(),
  features = NULL,
  layer = c("data", "scale.data", "counts"),
  na.rm = TRUE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{object}{An \code{\link[SeuratObject]{Assay}} object}

\item{data}{A \code{data.frame} (eg. from
\code{\link[SeuratObject]{FetchData}}) with extra data for visualization}

\item{features}{A vector of features to include in the fortified data;
defaults to the \link[SeuratObject:VariableFeatures]{variable features}}

\item{layer}{Name of layer to pull expression data for}

\item{na.rm}{Remove values with \code{NA}s}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{...}{Arguments passed to \code{\link[ggplot2]{geom_violin}}}
}
\value{
A \code{\link[ggplot2]{geom_violin}} layer to add to
\code{ggplot} objects
}
\description{
Generate a \code{\link[ggplot2]{geom_violin}} layer to add to
existing \code{ggplot} objects. The new plot data and aesthetics are stored
at the \code{geom} level
}
\details{
The following aesthetics are mapped automatically and overwrite
said aesthetics at the plot level if set:
\Sexpr[stage=build,results=rd]{ggseurat::.rd_ilist(c("x", "y", "fill"))}
}
\examples{
data("pbmc_small")
rna <- pbmc_small[["RNA"]]
md <- FetchData(pbmc_small, vars = "ident")

# Generate a plot using `autolayer()`

}
\seealso{
\code{\link[ggplot2:autolayer]{ggplot2::autolayer}()},
\code{\link[ggplot2:geom_violin]{ggplot2::geom_violin}()}

Visualize Assay Expression Data:
\code{\link{autoplot.Assay}()},
\code{\link{fortify.Assay}()}
}
\concept{assay}
