% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{autoplot.Assay}
\alias{autoplot.Assay}
\title{Create a Plot for \code{\link[SeuratObject]{Assay}} Objects}
\usage{
\method{autoplot}{Assay}(
  object,
  data = missing_arg(),
  features = NULL,
  layer = c("data", "scale.data", "counts"),
  na.rm = TRUE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{object}{An \code{\link[SeuratObject]{Assay}} object}

\item{data}{A \code{data.frame} (eg. from
\code{\link[SeuratObject]{FetchData}}) with extra data for visualization}

\item{features}{A vector of features to include in the fortified data;
defaults to the \link[SeuratObject:VariableFeatures]{variable features}}

\item{layer}{Name of layer to pull expression data for}

\item{na.rm}{Remove values with \code{NA}s}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{...}{Arguments passed to \code{\link[ggplot2]{geom_violin}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Generate a \code{ggplot} plot for
 \code{\link[SeuratObject]{Assay}} objects
}
\details{
The following aesthetics are mapped automatically:
\Sexpr[stage=build,results=rd]{ggseurat::.rd_ilist(c("x", "y", "fill"))}
}
\examples{
data("pbmc_small")
rna <- pbmc_small[["RNA"]]
md <- FetchData(pbmc_small, vars = "ident")

# Generate a plot using `autoplot()`

}
\seealso{
\code{\link[ggplot2:autoplot]{ggplot2::autoplot}()}

Visualize Assay Expression Data:
\code{\link{autolayer.Assay}()},
\code{\link{fortify.Assay}()}
}
\concept{assay}
