% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{autoplot.DimReduc}
\alias{autoplot.DimReduc}
\title{Create a Plot for \code{\link[SeuratObject]{DimReduc}} Objects}
\usage{
\method{autoplot}{DimReduc}(
  object,
  data = missing_arg(),
  dims = c(1L, 2L),
  na.rm = TRUE,
  show.legend = NA,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{DimReduc}} object}

\item{data}{A \code{data.frame} (eg. from
\code{\link[SeuratObject]{FetchData}}) with extra data for visualization}

\item{dims}{Dimensions for visualization}

\item{na.rm}{Remove values with \code{NA}s}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{...}{Arguments passed to \code{\link[ggplot2]{geom_point}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Generate a \code{ggplot} plot for
 \code{\link[SeuratObject]{DimReduc}} objects
}
\details{
The following aesthetics are mapped automatically:
\Sexpr[stage=build,results=rd]{ggseurat::.rd_ilist(c("x", "y", "color"))}
}
\examples{
data("pbmc_small")
tsne <- pbmc_small[["tsne"]]
md <- FetchData(pbmc_small, vars = c("ident", "MS4A1"))

# Generate a plot using `autoplot()`
autoplot(tsne)
autoplot(tsne, data = md)

}
\seealso{
\code{\link[ggplot2:autoplot]{ggplot2::autoplot}()}

Visualize Dimensional Reductions:
\code{\link{autolayer.DimReduc}()},
\code{\link{fortify.DimReduc}()}
}
\concept{dimreduc}
