% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.fortify}
\alias{.fortify}
\title{Prepare Fortified Data for `ggplot()`}
\usage{
.fortify(df, ident = TRUE, na.rm = FALSE)
}
\arguments{
\item{df}{A data frame with at least one column called \dQuote{\code{cell}}}

\item{ident}{If \code{TRUE}, add an identity column to \code{df}}

\item{na.rm}{Remove values with \code{NA}s}
}
\value{
...
}
\description{
Prepare Fortified Data for `ggplot()`
}
\keyword{internal}
