% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.get_aesthetics}
\alias{.get_aesthetics}
\title{Get Aesthetics From a Layer}
\usage{
.get_aesthetics(plot, layer, split = TRUE)
}
\arguments{
\item{plot}{A ggplot object}

\item{layer}{A ggplot layer}

\item{split}{Split the resulting vector into a list based on whether or not
the value is present in the layer data or the plot data}
}
\value{
A character vector of all aesthetics defined for the layer
}
\description{
Get Aesthetics From a Layer
}
\examples{
data("pbmc_small")
p <- ggplot(pbmc_small, mapping = aes(color = ident)) +
  geom_point(aes(x = tSNE_1, y = tSNE_2))
p

.get_aesthetics(p, p$layers[[1L]])

}
\keyword{internal}
