% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.oxford}
\alias{.oxford}
\title{Create a List with a Serial Comma}
\usage{
.oxford(
  ...,
  cnj = c("or", "and"),
  quote = c("single", "fancysingle", "double", "fancydouble", "none")
)
}
\arguments{
\item{...}{A character vector to join}

\item{cnj}{Conjunction to use for final entry}

\item{quote}{Quote the entries of \code{...}; choose from:
\itemize{
 \item \dQuote{\code{single}}: regular single quotes
 \item \dQuote{\code{fancysingle}}: fancy single quotes
 \item \dQuote{\code{double}}: regular double quotes
 \item \dQuote{\code{fancydouble}}: fancy double quotes
 \item \dQuote{\code{none}}: no extra quoting
}}
}
\value{
\code{...} arranged into an English list with a serial comma
when needed
}
\description{
Create a List with a Serial Comma
}
\examples{
.oxford('cell')
.oxford('cell', 'ident')
.oxford('cell', 'ident', 'gene')

}
\seealso{
\code{\link[base]{sQuote}}
}
\keyword{internal}
