% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.prep_plot_data}
\alias{.prep_plot_data}
\alias{.prep_plot_data.data.frame}
\alias{.prep_plot_data.default}
\alias{.prep_plot_data.list}
\alias{.prep_plot_data.NULL}
\title{Prep Data for Plotting}
\usage{
.prep_plot_data(data, idx, ...)

\method{.prep_plot_data}{data.frame}(data, idx, ...)

\method{.prep_plot_data}{default}(data, idx, name, restricted = "cell", ...)

\method{.prep_plot_data}{list}(data, idx, ...)

\method{.prep_plot_data}{`NULL`}(data, idx, ...)
}
\arguments{
\item{data}{Extra data for plotting}

\item{idx}{An \code{\link[base]{rle}} object where the values are the
possible index names and the lengths are the number of times those indexes
are present}

\item{...}{Arguments passed to other methods}

\item{name}{Name for resulting \code{\link[base]{data.frame}}}

\item{restricted}{A vector of names that \code{name} should not be}
}
\value{
...
}
\description{
Prep Data for Plotting
}
\examples{
data("pbmc_small")

}
\keyword{internal}
