% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{.rd_ident}
\alias{.rd_ident}
\title{Rdocumentation Default Identity}
\usage{
.rd_ident()
}
\value{
An Rdocumentation-formatted string containing the value of
\code{getOption("Seurat.object.project")}
}
\description{
Generate an Rdocumentation-formatted string with the default identity class
from \pkg{SeuratObject} as determined by
\code{getOption("Seurat.object.project")}; the default identity class is
double-quoted and in a monospace font (\code{\\dQuote{\\code{}}})
}
\section{Default Rendering Example}{

\Sexpr[stage=build,results=rd]{ggseurat::.rd_ident()}
}

\examples{
#' \Sexpr[results=rd]{ggseurat::.rd_ident()}
}
\keyword{documentation}
\keyword{internal}
