% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{.rd_ilist}
\alias{.rd_ilist}
\title{Rdocumentation Itemized List}
\usage{
.rd_ilist(x)
}
\arguments{
\item{x}{A vector of values for the list}
}
\value{
An Rdocumentation itemized list suitable for injecting into
documentation files with \code{\\Sexpr}
}
\description{
Generate an Rdocumentation itemized list
}
\examples{
#' \Sexpr[results=rd]{ggseurat::.rd_ilist()}
}
\keyword{documentation}
\keyword{internal}
