% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay.R
\name{fortify.Assay}
\alias{fortify.Assay}
\title{Create a Data Frame Suitable for \code{ggplot()}}
\usage{
\method{fortify}{Assay}(
  model,
  data,
  features = NULL,
  layer = c("data", "scale.data", "counts"),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{model}{An \code{\link[SeuratObject]{Assay}} object}

\item{data}{A \code{data.frame} (eg. from
\code{\link[SeuratObject]{FetchData}}) with extra data for visualization}

\item{features}{A vector of features to include in the fortified data;
defaults to the \link[SeuratObject:VariableFeatures]{variable features}}

\item{layer}{Name of layer to pull expression data for}

\item{na.rm}{Remove values with \code{NA}s}

\item{...}{Ignored}
}
\value{
A long-format data frame for use with \code{\link[ggplot2]{ggplot}()}
}
\description{
Generate a long-format data frame for ease of use with
\code{\link[ggplot2]{ggplot}()}; as a bonus, this method enables direct
calling of \code{\link[SeuratObject]{Assay}} objects in
\code{\link[ggplot2]{ggplot}()} (see examples)
}
\details{
\code{fortify.Assay()} generates a data frame based on the
expression data stored in \code{layer}; automatically creates a column for
cell names (\dQuote{\code{cell}}). Also creates creates a column for
identity classes (\dQuote{\code{ident}}) if not present in \code{data}; the
default identity class is
\Sexpr[stage=render, results=rd]{ggseurat::.rd_ident()}; final column output
order is:
\itemize{
 \item \dQuote{\code{cell}}
 \item \dQuote{\code{ident}}
 \item expression data for \code{features}
 \item additional meta data provided by \code{data}
}
}
\examples{
data("pbmc_small")
rna <- pbmc_small[["RNA"]]
md <- FetchData(pbmc_small, vars = "ident")

# Create a data frame for `ggplot()`
df <- fortify(rna, data = md)
head(df)
ggplot(df, mapping = aes(x = ident, y = PPBP, fill = ident)) +
  geom_violin()

# Use an `Assay` directly in `ggplot()`
ggplot(rna, mapping = aes(x = ident, y = GNLY, fill = ident), md) +
  geom_violin()

}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot}()},
\code{\link[ggplot2:fortify]{ggplot2::fortify}()}

Visualize Assay Expression Data:
\code{\link{autolayer.Assay}()},
\code{\link{autoplot.Assay}()}
}
\concept{assay}
