% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduc.R
\name{fortify.DimReduc}
\alias{fortify.DimReduc}
\title{Create a Data Frame Suitable for \code{ggplot()}}
\usage{
\method{fortify}{DimReduc}(model, data, na.rm = FALSE, ...)
}
\arguments{
\item{model}{A \code{\link[SeuratObject]{DimReduc}} object}

\item{data}{A \code{data.frame} (eg. from
\code{\link[SeuratObject]{FetchData}}) with extra data for visualization}

\item{na.rm}{Remove values with \code{NA}s}

\item{...}{Ignored}
}
\value{
A long-format data frame for use with \code{\link[ggplot2]{ggplot}()}
}
\description{
Generate a long-format data frame for ease of use with
\code{\link[ggplot2]{ggplot}()}; as a bonus, this method enables direct
calling of \code{\link[SeuratObject]{DimReduc}} objects in
\code{\link[ggplot2]{ggplot}()} (see examples)
}
\details{
\code{fortify.DimReduc()} generates a data frame based on the
\link[SeuratObject:Embeddings]{cell embeddings}. Automatically creates a
column for cell names (\dQuote{\code{cell}}). Also creates a column for
identity classes (\dQuote{\code{ident}}) if not present in \code{data}; the
default identity class is
\Sexpr[stage=render, results=rd]{ggseurat::.rd_ident()}; final column output
order is:
\itemize{
 \item \dQuote{\code{cell}}
 \item \dQuote{\code{ident}}
 \item cell embeddings
 \item additional meta data provided by \code{data}
}
}
\examples{
data("pbmc_small")
tsne <- pbmc_small[["tsne"]]
md <- FetchData(pbmc_small, vars = c("ident", "MS4A1"))

# Create a data frame for `ggplot()`
df <- fortify(tsne, data = md)
head(df)
ggplot(df, mapping = aes(x = tSNE_1, y = tSNE_2, color = ident)) +
  geom_point()

# Use a `DimReduc` directly in `ggplot()`
ggplot(tsne, mapping = aes(x = tSNE_1, y = tSNE_2, color = MS4A1), md) +
  geom_point()

}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot}()},
\code{\link[ggplot2:fortify]{ggplot2::fortify}()}

Visualize Dimensional Reductions:
\code{\link{autolayer.DimReduc}()},
\code{\link{autoplot.DimReduc}()}
}
\concept{dimreduc}
