% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{GetAesthetics}
\alias{GetAesthetics}
\title{Get aesthetics from a layer}
\usage{
GetAesthetics(plot, layer, split = TRUE)
}
\arguments{
\item{plot}{A ggplot object}

\item{layer}{A ggplot layer}

\item{split}{Split the resulting vector into a list based on whether or not
the value is present in the layer data or the plot data}
}
\value{
A character vector of all aesthetics defined for the layer
}
\description{
Get aesthetics from a layer
}
\keyword{internal}
