% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot.Seurat}
\alias{ggplot.Seurat}
\alias{ggplot}
\title{Visualize Seurat objects}
\usage{
\method{ggplot}{Seurat}(data = NULL, mapping = aes(), ...,
  environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[=fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{environment}{DEPRECATED. Used prior to tidy evaluation.}
}
\value{
A ggplot object
}
\description{
Visualize Seurat objects
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
