% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_utils.R
\docType{data}
\name{ggplot2-utils}
\alias{ggplot2-utils}
\alias{create_layout}
\alias{empty}
\alias{find_global}
\alias{make_labels}
\alias{match_calculated_aes}
\alias{message_wrap}
\alias{new_data_frame}
\alias{plot_clone}
\alias{scales_add_missing}
\alias{ScalesList}
\alias{scales_list}
\alias{scales_map_df}
\alias{scales_train_df}
\alias{scales_transform_df}
\alias{strip_dots}
\title{ggplot2 utility functions}
\format{An object of class \code{character} of length 1.}
\usage{
create_layout(facet = FacetNull, coord = CoordCartesian)

empty(df)

find_global(name, env, mode = "any")

make_labels(mapping)

match_calculated_aes

message_wrap(...)

new_data_frame(x = list(), n = NULL)

plot_clone(plot)

scales_add_missing(plot, aesthetics, env)

ScalesList

scales_list()

scales_map_df(scales, df)

scales_train_df(scales, df, drop = FALSE)

scales_transform_df(scales, df)

strip_dots(expr)
}
\description{
These functions are internal functions copied from ggplot2 to maintain
functionality. All credit goes to the ggplot2 authors
}
\seealso{
\code{\link[ggplot2]{ggplot2-package}}
}
\author{
ggplot2 developers
}
\keyword{datasets}
\keyword{internal}
