% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_build.ggseurat}
\alias{ggplot_build.ggseurat}
\alias{ggplot_build}
\title{ggplot_build.Seurat}
\usage{
\method{ggplot_build}{ggseurat}(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\value{
A built ggplot ready for viewing
}
\description{
ggplot_build.Seurat
}
\seealso{
\code{\link[ggplot2]{ggplot_build}}
}
