% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ggseurat-package}
\alias{ggseurat}
\alias{ggseurat-package}
\title{ggseurat: ggplot2 Bindings for Seurat Objects}
\description{
Provides methods to allow Seurat objects to be utilized directly in
    the ggplot2 ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mojaveazure/ggseurat}
  \item Report bugs at \url{https://github.com/mojaveazure/ggseurat/issues}
}

}
\author{
\strong{Maintainer}: Paul Hoffman \email{phoffman@nygenome.org} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

}
